+++ CHANGES.rst
@@ -4,10 +4,13 @@ Releases
 0.1.0 (2013-11-05)
 ------------------
 
+* Fix a bug when you deserializer and the format to the fixtures was XML
 * Add message when raise exceptions
 * Improvements in the serialization form in the example project
 * Fix a little error when you serialized and the format was python
 * Reorder and refactor the code. **There are backward incompatible changes** like reorder of the params of the methods
+* Improvements in the tests
+
 
 
 0.0.1 (2013-10-29)